# packages/server/api/v1.swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API for Chalarí",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/authentication/register": {
            "summary": "Registers a new user in the database.",
            "post": {
                "tags": ["Authentication"],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "firstName": {
                                        "type": "string",
                                        "example": "John"
                                    },
                                    "lastName": {
                                        "type": "string",
                                        "example": "Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "john.doe@chalari.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "johndoe01"
                                    }
                                },
                                "required": ["firstName", "lastName", "email", "password"]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Registered the user successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Registered the user successfully"
                                        }
                                    },
                                    "required": ["success", "message"]
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict. The provided email is already being used.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "A user with the given email already exists."
                                        }
                                    },
                                    "required": ["success", "message"]
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "The body format doesn't match the expected format.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Bad request format."
                                        }
                                    },
                                    "required": ["success", "message"]
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Something went wrong whilst registering a user"
                                        }
                                    },
                                    "required": ["success", "message"]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/authentication/login": {
            "summary": "Logs a new user.",
            "post": {
                "tags": ["Authentication"],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "john.doe@chalari.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "johndoe01"
                                    }
                                },
                                "required": ["email", "password"]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Logged the user in successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Logged the user in successfully"
                                        }
                                    },
                                    "required": ["success", "message"]
                                }
                            }
                        },
                        "headers": {
                            "Set-Cookie": {
                                "description": "The JWT token to access protected resources.",
                                "schema": {
                                    "type": "string",
                                    "example": "34209rh32-12309hf3bnr-495u2394th-34t43295"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Invalid credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid credentials."
                                        }
                                    },
                                    "required": ["success", "message"]
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "The body format doesn't match the expected format.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Bad request format."
                                        }
                                    },
                                    "required": ["success", "message"]
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Something went wrong whilst logging a user in"
                                        }
                                    },
                                    "required": ["success", "message"]
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}